refactor(pack): refactor xdsl dmz
